all necessary code added

+++ app/Http/Controllers/Resource/GdnDetailController.php
@@ -18,9 +18,9 @@ class GdnDetailController extends Controller
 
         abort_if($gdnDetail->gdn->reservation()->exists(), 403);
 
-        abort_if($gdnDetail->gdn->isSubtracted(), 403);
+        abort_if($gdnDetail->gdn->isSubtracted() || $gdnDetail->gdn->isCancelled(), 403);
 
-        abort_if($gdnDetail->gdn->isApproved() && ! authUser()->can('Delete Approved GDN'), 403);
+        abort_if($gdnDetail->gdn->isApproved() && !authUser()->can('Delete Approved GDN'), 403);
 
         $gdnDetail->forceDelete();
